refactor(wallet-core): extract persistentDeviceDataReducer as a subreducer
