+++ suite-common/device/src/index.ts
@@ -2,6 +2,9 @@ export * from './deviceActions';
 export * from './deviceConstants';
 export * from './deviceReducer';
 export * from './deviceSelectors';
+export * from './persistentDeviceDataActions';
+export * from './persistentDeviceDataReducer';
+export * from './persistentDeviceDataSelectors';
 export * from './deviceUtils';
 export * from './sortDevices';
 export * from './usePinHook';
